import changes from VDB-3815

+++ libs/sra/reader-abi.c
@@ -283,7 +283,7 @@ LIB_EXPORT rc_t CC AbsolidReaderHeader(const AbsolidReader* self, uint32_t readI
     if( self->dad.options & eOrigFormat ) {
         char tmp[1024];
         if( spotname_sz == 0 ) {
-            spotname_sz = snprintf(tmp, sizeof(tmp) - 1, "%s.%lld", self->dad.accession, self->dad.spot);
+            spotname_sz = snprintf(tmp, sizeof(tmp) - 1, "%s.%lld", self->dad.accession, ( long long int ) self->dad.spot);
             if ( spotname_sz < 0 )
                 return RC ( rcSRA, rcString, rcConstructing, rcData, rcCorrupt );
             spotname = tmp;
@@ -296,10 +296,10 @@ LIB_EXPORT rc_t CC AbsolidReaderHeader(const AbsolidReader* self, uint32_t readI
     } else {
         if( readId > 0 && read_label_sz > 0 ) {
             ret = snprintf(data, dsize, ">%s.%lld %.*s%s%.*s",
-                self->dad.accession, self->dad.spot, spotname_sz, spotname, spotname_sz ? "_" : "", read_label_sz, read_label);
+                self->dad.accession, ( long long int ) self->dad.spot, spotname_sz, spotname, spotname_sz ? "_" : "", read_label_sz, read_label);
         } else {
             ret = snprintf(data, dsize, ">%s.%lld %.*s",
-                           self->dad.accession, self->dad.spot, spotname_sz, spotname);
+                           self->dad.accession, ( long long int ) self->dad.spot, spotname_sz, spotname);
         }
     }
 #if SNPRINTF_ACTUALLY_WORKED_THE_WAY_YOU_THINK
